Bash Script to stop CrashLoops - Airflow Scheduler
The Airflow scheduler kills itself after scheduling a
dag run.  This will cause the Airflow Scheduler container
to go into 'CrashLoopBackOff' after a while.  This is very
disruptive to our workflow as it takes a long time for
the container to get back into 'Running' state (Airflow
will not be able to schedule and perform other tasks while
the scheduler is down).

The current workaround is to run a bash script that will kick
start the airflow scheduler process whenever it goes down.  In
this way we will be able to avoid the Airflow Scheduler container
from crashing.

Change-Id: I6c0e4277d269654692e6577ef849e98c3ab886d4