Update patch set 1
Patch Set 1: Code-Review-1

If we are going to start using this repository again. We need to update the Makefile to work the same as Airship components.  Labels need to be added, the Jenkins image we produce needs to have the SHA that it pulled from.

Patch-set: 1
Reviewer: Gerrit User 1011593 <1011593@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1