Add retries to stack delete/create in shaker spot
Rather than simply waiting for the stack status to get to the point
where we want it to be, retry the action we performed after a given
set of time. For delete, if we've waited for the delete to complete
for awhile, issue another delete to see if things behave better.
For create, if we've waited for the create to complete for awhile,
attempte a delete/create.

Change-Id: I3f89f678097c243c2c715b8a61e9ea86cf77c1ea